+++ model/src/main/scala/aqua/model/inline/state/Arrows.scala
@@ -1,15 +1,26 @@
-package aqua.model.func
+package aqua.model.inline.state
 
 import aqua.raw.arrow.{ArgsCall, FuncArrow, FuncRaw}
 import cats.data.State
-import cats.syntax.traverse.*
-import cats.syntax.functor.*
 import cats.instances.list.*
+import cats.syntax.functor.*
+import cats.syntax.traverse.*
 
+/**
+ * State algebra for resolved arrows
+ *
+ * @tparam S State
+ */
 trait Arrows[S] extends Scoped[S] {
   self =>
   def save(name: String, arrow: FuncArrow): State[S, Unit]
 
+  /**
+   * Arrow is resolved – save it to the state [[S]]
+   *
+   * @param arrow resolved arrow
+   * @param e     contextual Exports that an arrow captures
+   */
   final def resolved(arrow: FuncRaw)(implicit e: Exports[S]): State[S, Unit] =
     for {
       exps <- e.exports
@@ -18,17 +29,43 @@ trait Arrows[S] extends Scoped[S] {
       _ <- save(arrow.name, funcArrow)
     } yield ()
 
+  /**
+   * Save arrows to the state [[S]]
+   *
+   * @param arrows Resolved arrows, accessible by key name which could differ from arrow's name
+   */
   final def resolved(arrows: Map[String, FuncArrow]): State[S, Unit] =
     arrows.toList.traverse(save).void
 
+  /**
+   * All arrows available for use in scope
+   */
   val arrows: State[S, Map[String, FuncArrow]]
 
+  /**
+   * Pick a subset of arrows by names
+   *
+   * @param names What arrows should be taken
+   */
   def pickArrows(names: Set[String]): State[S, Map[String, FuncArrow]] =
     arrows.map(_.view.filterKeys(names).toMap)
 
+  /**
+   * Take arrows selected by the function call arguments
+   *
+   * @param args
+   * @return
+   */
   def argsArrows(args: ArgsCall): State[S, Map[String, FuncArrow]] =
     arrows.map(args.arrowArgs)
 
+  /**
+   * Changes the [[S]] type to [[R]]
+   *
+   * @param f Lens getter
+   * @param g Lens setter
+   * @tparam R New state type
+   */
   def transformS[R](f: R => S, g: (R, S) => R): Arrows[R] = new Arrows[R] {
 
     override def save(name: String, arrow: FuncArrow): State[R, Unit] =
@@ -48,6 +85,7 @@ trait Arrows[S] extends Scoped[S] {
 object Arrows {
   def apply[S](implicit arrows: Arrows[S]): Arrows[S] = arrows
 
+  // Default implementation with the most straightforward state – just a Map
   object Simple extends Arrows[Map[String, FuncArrow]] {
 
     override def save(name: String, arrow: FuncArrow): State[Map[String, FuncArrow], Unit] =
