+++ model/src/main/scala/aqua/model/inline/state/Counter.scala
@@ -1,14 +1,25 @@
-package aqua.model.func
+package aqua.model.inline.state
 
 import cats.data.State
 import cats.syntax.flatMap.*
 
+/**
+ * Monotonic counter, stored within the State monad
+ *
+ * @tparam S State
+ */
 trait Counter[S] {
   self =>
+  // Get counter
   val get: State[S, Int]
+
+  // Increment by i
   def add(i: Int): State[S, Unit]
+
+  // Increment by 1 and get
   val incr: State[S, Int] = add(1) >> get
 
+  // Change state [[S]] to [[R]]
   def transformS[R](f: R => S, g: (R, S) => R): Counter[R] = new Counter[R] {
     override val get: State[R, Int] = self.get.transformS(f, g)
 
@@ -22,6 +33,7 @@ object Counter {
   object Simple extends Counter[Int] {
 
     override val get: State[Int, Int] = State.get
+
     override def add(i: Int): State[Int, Unit] = State.modify[Int](_ + i)
   }
 }
