+++ model/src/main/scala/aqua/model/inline/state/InliningState.scala
@@ -1,0 +1,42 @@
+package aqua.model.inline.state
+
+import aqua.model.ValueModel
+import aqua.model.inline.state.{Arrows, Counter, Exports, Mangler}
+import aqua.raw.arrow.{FuncArrow, FuncRaw}
+import aqua.raw.value.{ValueRaw, VarRaw}
+import aqua.types.ArrowType
+import cats.data.{Chain, State}
+import cats.instances.list.*
+import cats.syntax.traverse.*
+import scribe.Logging
+
+/**
+ * Default states aggregate for all available state algebras
+ *
+ * @param noNames            for [[Mangler]]
+ * @param resolvedExports    for [[Exports]]
+ * @param resolvedArrows     for [[Arrows]]
+ * @param instructionCounter for [[Counter]]
+ */
+case class InliningState(
+                          noNames: Set[String] = Set.empty,
+                          resolvedExports: Map[String, ValueRaw] = Map.empty,
+                          resolvedArrows: Map[String, FuncArrow] = Map.empty,
+                          instructionCounter: Int = 0
+                        )
+
+object InliningState {
+
+  given Counter[InliningState] =
+    Counter.Simple.transformS(_.instructionCounter, (acc, i) => acc.copy(instructionCounter = i))
+
+  given Mangler[InliningState] =
+    Mangler.Simple.transformS(_.noNames, (acc, nn) => acc.copy(noNames = nn))
+
+  given Arrows[InliningState] =
+    Arrows.Simple.transformS(_.resolvedArrows, (acc, aa) => acc.copy(resolvedArrows = aa))
+
+  given Exports[InliningState] =
+    Exports.Simple.transformS(_.resolvedExports, (acc, ex) => acc.copy(resolvedExports = ex))
+
+}
