+++ model/src/main/scala/aqua/model/inline/state/Mangler.scala
@@ -1,4 +1,4 @@
-package aqua.model.func
+package aqua.model.inline.state
 
 import cats.data.State
